+++ src/app/top-toolbar/top-toolbar.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, DoCheck, ViewChild} from '@angular/core';
+import {Component, DoCheck, ViewChild} from '@angular/core';
 import {AuthenticationService} from '../../services/authentication.service';
 import {Router} from '@angular/router';
 import {PalladiumApiService} from '../../services/palladium-api.service';
@@ -6,19 +6,16 @@ import {PalladiumApiService} from '../../services/palladium-api.service';
 @Component({
   selector: 'app-top-toolbar',
   templateUrl: './top-toolbar.component.html',
-  styleUrls: ['./top-toolbar.component.css'],
+  styleUrls: ['./top-toolbar.component.scss'],
   providers: [PalladiumApiService]
 })
-export class TopToolbarComponent implements OnInit, DoCheck {
+export class TopToolbarComponent implements DoCheck {
   public authorize;
   public statuses = {};
   @ViewChild('Modal') Modal;
 
   constructor(private router: Router,
               private authenticationService: AuthenticationService) {
-  }
-
-  ngOnInit() {
   }
 
   ngDoCheck() {