+++ Changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 ## master (unreleased)
-
+## [0.4.6] - 2018-11-27
+## Added
+    * Change bootstrap modals to angular material dialogs in token settings
+## Changed
+    * Get invite and generate invite methods is changed for work with await/async
+    * Invite modal is rework to angular material dialog
+    * About modal is rework to angular material dialog
+    * Status setting modal is rework to angular material dialog
+    * Delete unused scss file from invite component
 ## [0.4.5] - 2018-06-15
 ## Fixed
     * Fix opening add_result menu without result_set select